Added several linked lists functions to work on the push_swap project. Implementing find median with median of medians algorithm

+++ src/linked_lists/ft_lst_find_max.c
@@ -6,21 +6,21 @@
 /*   By: dda-silv <dda-silv@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/11 10:03:33 by dda-silv          #+#    #+#             */
-/*   Updated: 2021/04/11 10:10:55 by dda-silv         ###   ########.fr       */
+/*   Updated: 2021/04/11 16:14:31 by dda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_lst_find_max(t_list *lst)
+long long	ft_lst_find_max(t_list *lst)
 {
-	int	max;
+	long long	max;
 
-	max = (long int)lst->data;
+	max = (long long int)lst->data;
 	while (lst)
 	{
-		if (max < (long int)lst->data)
-			max = (long int)lst->data;
+		if (max < (long long int)lst->data)
+			max = (long long int)lst->data;
 		lst = lst->next;
 	}
 	return (max);
