+++ src/linked_lists/ft_lst_find_min.c
@@ -6,21 +6,21 @@
 /*   By: dda-silv <dda-silv@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/11 10:06:59 by dda-silv          #+#    #+#             */
-/*   Updated: 2021/04/11 10:10:54 by dda-silv         ###   ########.fr       */
+/*   Updated: 2021/04/11 16:14:23 by dda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_lst_find_min(t_list *lst)
+long long	ft_lst_find_min(t_list *lst)
 {
-	int	min;
+	long long	min;
 
-	min = (long int)lst->data;
+	min = (long long int)lst->data;
 	while (lst)
 	{
-		if ((long int)lst->data < min)
-			min = (long int)lst->data;
+		if ((long long int)lst->data < min)
+			min = (long long int)lst->data;
 		lst = lst->next;
 	}
 	return (min);
